Checks for missing and gap-filled data and determines output years

Checks for missing and gap-filled data and determines output years

CheckDataGaps(
  datain,
  qc_flags,
  missing_met,
  missing_flux,
  gapfill_met_tier1,
  gapfill_met_tier2,
  gapfill_flux,
  gapfill_good,
  gapfill_med,
  gapfill_poor,
  gapfill_era,
  gapfill_stat,
  min_yrs,
  qc_name,
  aggregate = NA,
  site_log
)

Arguments

datain

Input data list

qc_flags

Values of qc flags

missing_met

Threshold for missing values per year for met variables (as percentage)

missing_flux

Threshold for missing values per year for flux variables (as percentage)

gapfill_met_tier1

Threshold for all gap_filling per year for Tier 1 met variables (as percentage)

gapfill_met_tier2

Threshold for all gap_filling per year for Tier 2 met variables (as percentage)

gapfill_flux

Threshold for all gap_filling per year for flux variables (as percentage)

gapfill_good

Threshold for good quality gap_filling per year (as percentage, ignored if gapfill_all set)

gapfill_med

Threshold for medium quality gap_filling per year (as percentage, ignored if gapfill_all set)

gapfill_poor

Threshold for poor quality gap_filling per year (as percentage, ignored if gapfill_all set)

min_yrs

Minimum number of consecutive years

qc_name

Name of QC variable of dataset

site_log

Site log

showWarn

Print warning?

Value

out