Default options for a dataset conversion

Default options for a dataset conversion

get_default_conversion_options()

Value

options list, with values:

datasetname: Name of the dataset, e.g. FLUXNET2015 or La Thuile. Defaults to FLUXNET2015, and thus must be set if processing a dataset not compliant with FLUXNET2015 format.

datasetversion: Version of the dataset, e.g. "1-3"

flx2015_version: Version of FLUXNET2015 data product being used, i.e. "FULLSET" or "SUBSET". Required to set QC flags correctly.

fair_use: La Thuile Fair Use policy that data should comply with, e.g. "LaThuile" or "Fair_Use" (default). Can be a single entry or a vector of several policies. If this is set, code will only extract years that comply with the required policy/policies. Must provide fair_use_vec to use this functionality.

fair_use_vec: A vector of Data Use policy for each year in the data file, e.g. "LaThuile" or "Fair_Use". Should have years as vector column names.

met_gapfill: Method to use for gap-filling meteorological data. Set to one of "ERAinterim", "statistical" or NA (default; no gap-filling).

flux_gapfill: Method to use for gap-filling flux data. Set to one of "statistical" or NA (default; no gap-filling).

missing: Maximum percentage of time steps allowed to be missing in any given year

gapfill_met_tier1: Maximum percentage of time steps allowed to be gap-filled (any quality) in any given year in Tier 1 met variables. Set to NA if not required.

gapfill_met_tier2: Maximum percentage of time steps allowed to be gap-filled (any quality) in any given year in Tier 2 met variables. Set to NA if not required.

gapfill_flux: Maximum percentage of time steps allowed to be gap-filled (any quality) in any given year in flux variables. Set to NA if not required.

gapfill_good: Maximum percentage of time steps allowed to be good-quality gap-filled in any given year. Refer to package documentation for information on QC flags. Set to NA if not required (default).

gapfill_med: Maximum percentage of time steps allowed to be medium-quality gap-filled in any given year. Refer to package documentation for information on QC flags. Set to NA if not required (default).

gapfill_poor: Maximum percentage of time steps allowed to be poor-quality gap-filled in any given year. Refer to package documentation for information on QC flags. Set to NA if not required (default).

gapfill_era: Maximum percentage of time steps allowed to be ERA-Interim gap-filled in any given year. Refer to package documentation for information on QC flags. Set to NA if not required (default).

gapfill_stat: Maximum percentage of time steps allowed to be statistically gap-filled in any given year. Refer to package documentation for information on QC flags. Set to NA if not required (default).

min_yrs: Minimum number of consecutive years to process

linfill: Maximum consecutive length of time (in hours) to be gap-filled using linear interpolation. Used for all variables except rainfall. Defaults to 4 hours.

copyfill: Maximum consecutive length of time (in number of days) to be gap-filled using copyfill. Defaults to 10 days.

regfill: Maximum consecutive length of time (in number of days) to be gap-filled using multiple linear regression. Defaults to 30 days. Default method used to gapfill flux variables. If gapfilling by copyfill is preferred, set regfill to NA.

lwdown_method: Method used to synthesize incoming longwave radiation. One of "Abramowitz_2012" (default), "Swinbank_1963" or "Brutsaert_1975".

check_range_action: Action to take when input data falls outside of valid ranges (as defined in data/Output_variables_*.csv). One of "stop" (log error and stop processing, default), "warn" (log error and continue), "ignore" (continue), or "truncate" (set values outside the valid range to the range bounds).

include_all_eval: Should all evaluation values be included, regardless of data gaps? If set to FALSE, any evaluation variables with missing or gap-filled values in excess of the thresholds will be discarded.

aggregate: Time step (in hours) that the data is aggregated to. Must be divisible by 24 and can be set to a maximum 24 hours (daily). Defaults to NA (no aggregation).

model: Name of land surface model. Used to retrieve model specific attributes, such as site plant functional type.

metadata_source: Sources to check for metadata. One of 'all', 'csv', or 'web'.

add_psurf: La Thuile dataset does not include surface air pressure (PSurf). When set to TRUE, it will be synthesised from air temperature and elevation.